requirements-minimal.txt
# Requirements minimal pour SmartAgriDecision
# Installation rapide pour développement IA

# === CORE ESSENTIALS ===
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# === DATABASE ===
psycopg2-binary>=2.9.0

# === WEATHER DATA ===
meteostat>=1.0.0
requests>=2.31.0

# === GEOSPATIAL (optionnel) ===
geopandas>=0.13.0
shapely>=2.0.0

# === VISUALIZATION ===
matplotlib>=3.7.0
plotly>=5.15.0

# === UTILITIES ===
python-dateutil>=2.8.0
tqdm>=4.65.0

# === TESTING ===
pytest>=7.4.0
